Move UpdateConnectorHelper to utils package
